Split reservoir.rs (1001 lines) into reservoir/ subdirectory
Split into 5 files under reservoir/:
- mod.rs (12 lines): module declarations and re-exports
- types.rs (128 lines): constants, enums, structs
- systems.rs (157 lines): update_reservoir_levels system + ReservoirPlugin
- tests_unit.rs (437 lines): unit tests for types, constants, calculations
- tests_simulation.rs (287 lines): integration-style simulate_tick tests

All public items re-exported from mod.rs so external imports unchanged.
